// apps/web/src/types.ts
export type IconName =
  | 'activity'
  | 'alert'
  | 'architecture'
  | 'arrow'
  | 'asset'
  | 'check'
  | 'chevron'
  | 'clock'
  | 'close'
  | 'control'
  | 'database'
  | 'download'
  | 'filter'
  | 'globe'
  | 'grid'
  | 'history'
  | 'layers'
  | 'menu'
  | 'minus'
  | 'more'
  | 'plus'
  | 'report'
  | 'risk'
  | 'scan'
  | 'search'
  | 'server'
  | 'settings'
  | 'shield'
  | 'sparkles'
  | 'threat'
  | 'user'
  | 'vulnerability'
  | 'zoom-in'
  | 'zoom-out';